// client/src/generated/AuthResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { User } from "./User";

export type AuthResponse = { access_token: string, refresh_token: string, 
/**
 * Access-token lifetime in seconds.
 */
expires_in: number, user: User, };