// client/src/generated/Invite.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { UserId } from "./UserId";

export type Invite = { code: string, created_by: UserId, expires_at: number | null, max_uses: number | null, uses: number, revoked_at: number | null, created_at: number, };